commhandler/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "SGrPythontks4r"
dynamic = ["version"]
description = "SmartGridready CommHandler Library"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "BSD" }
authors = [{ name = "Robin Schoch" }]
urls = { "Homepage" = "https://github.com/SmartGridready/SGrPython" }
dependencies = [
    "Jinja2>=3.0.0,<4.0.0",
    "jmespath>=1.0.0,<2.0.0",
    "pymodbus>=3.0.0,<4.0.0",
    "setuptools>=68.0.0,<70.0.0",
    "xsdata>=22.0.0,<23.0.0",
    "aiohttp>=3.0.0,<4.0.0",
    "certifi",
    "cachetools>=5.0.0,<6.0.0",
    "SGrSpecificationPythontks4r==0.3.0",
]
[tool.pyright]
exclude = ["**/node_modules", "**/__pycache__", "example"]

[tool.pytest]
pythonpath = 'src'
asyncio_default_fixture_loop_scope = 'session'
asyncio_mode = 'strict'
